move getDisplayName to locale
