+++ pages/api/lib/_helpers.ts
@@ -181,15 +181,3 @@ export const getSubmissionDates = () => {
 
   return { previousSubmitTimestamp, nextSubmitTimestamp };
 };
-
-export const getDisplayName = (user: Member) => {
-  const { first_name, last_name, username } = user.about;
-  if (username) {
-    return username;
-  }
-  const userFullName = `${first_name || ''} ${last_name || ''}`.trim();
-  if (userFullName !== '') {
-    return userFullName;
-  }
-  return 'Anonymous';
-}
